route catalog authenticate ok!

+++ src/controllers/catalogController.js
@@ -1,9 +1,21 @@
 import catalogModel from "../models/catalogModel.js";
+// import { authUserAdm } from "../middlewares/authMiddleware.js";
 
 const catalogController = {
 
-    async getAllCatalog(req, res) {
+    async getAllCatalog(req, res) { // Lista todos o catalogo OK! USER
         try {
+
+            const creatorUserType = req.createUserType;
+            console.log("usuario autenticado: ",creatorUserType);
+    
+            // Impedindo usuários não-usuarios da PobreFlix de acessar todo o catálogo
+            if (creatorUserType !== ("Administrator" || "Client")) {
+                return res.status(403).json({
+                    error: "Só usuário da PobreFlix podem visualizar todo o Catálogo.",
+                });
+            }
+
             if (Object.keys(req.query).length > 0) {
                 return res.status(400).json({ 
                     error: "Esta rota não aceita parâmetros. Use /catalog/:title para consultas específicas." 
@@ -17,8 +29,19 @@ const catalogController = {
     },
     
     // Rota para consultar filmes por tipo
-    async getCatalogByType(req, res){
+    async getCatalogByType(req, res){ // Lista todos pelo tipo Catálogo OK! USER
         try {
+
+            const creatorUserType = req.createUserType;
+            console.log("usuario autenticado: ",creatorUserType);
+    
+            // Impedindo usuários não-usuarios da PobreFlix de acessar o 
+            if (creatorUserType !== ("Administrator" || "Client")) {
+                return res.status(403).json({
+                    error: "Só usuário da PobreFlix podem visualizar o Catálogo.",
+                });
+            }
+
             if (Object.keys(req.query).length > 0) {
                 return res.status(400).json({
                     error: "Esta rota não aceita parâmetros. Use /catalog/:title para consultas específicas.",
@@ -32,8 +55,20 @@ const catalogController = {
     },
     
     // Rota para consultar filme pelo título
-    async getCatalogByTitle(req, res) {
+    async getCatalogByTitle(req, res) { // Lista Todos pelo title Catálogo OK! USER
         try {
+            
+            const creatorUserType = req.createUserType;
+            console.log("usuario autenticado: ",creatorUserType);
+    
+            // Impedindo usuários não-usuarios da PobreFlix de acessar o titulo especifico
+            if (creatorUserType !== ("Administrator" || "Client")) {
+                return res.status(403).json({
+                    error: "Só usuário da PobreFlix podem visualizar o titulo do Catálogo.",
+                });
+            }
+
+
             if (Object.keys(req.query).length > 0) {
                 return res.status(400).json({
                     error: "Esta rota não aceita parâmetros. Use /catalog/:title para consultas específicas.",
@@ -47,10 +82,14 @@ const catalogController = {
     },
     
     // Rota para adicionar um filme ao catálogo
-    async createCatalog(req, res){
+    async createCatalog(req, res){ // Adicionar filme no catalogo OK! ADM
         try {
+
+             // Tipo padrão: todos os usuários comuns são criados como Client
+            const creatorUserType = req.createUserType;
+
             const { title, description, genre, content_type, video_url } = req.body;
-            const movie = await catalogModel.addToCatalog(title, description, genre, content_type, video_url);
+            const movie = await catalogModel.addToCatalog(title, description, genre, content_type, video_url, creatorUserType);
             res.status(201).json(movie);
         } catch (err) {
             res.status(500).json({ error: `Erro ao cadastrar filme: ${err.message}` });
@@ -58,11 +97,22 @@ const catalogController = {
     },
     
     // Rota para atualizar um filme no catálogo
-    async upCatalog (req, res) {
+    async upCatalog (req, res) { // Atualizar filme no catalogo OK! ADM
         try {
+
             const { id } = req.params;
             const { title, description, genre, content_type, video_url } = req.body;
+
+            const creatorUserType = req.createUserType;
+            console.log(`Usuário autenticado: ${creatorUserType}`);
     
+            // validações de atualização
+            if (creatorUserType !== 'Administrator') {
+                return res.status(403).json({
+                    error: "Apenas administradores podem alterar o Catalógo.",
+                });
+            }
+
             if (!title && !description && !genre && !content_type && !video_url) {
                 return res.status(400).json({ error: "Pelo menos um campo deve ser fornecido para atualização." });
             }
@@ -71,6 +121,7 @@ const catalogController = {
                 return res.status(400).json({ error: "O campo content_type deve ser 'filme' ou 'serie'." });
             }
     
+            // consulta ao banco e atualização
             const movie = await catalogModel.updateCatalog(id, { title, description, genre, content_type, video_url });
     
             if (!movie) {
@@ -84,10 +135,18 @@ const catalogController = {
     },
     
     // Rota para excluir um filme do catálogo
-    async deleteCatalog(req, res) {
+    async deleteCatalog(req, res) { // Excluir filme no catalogo OK! ADM
         try {
             const { id } = req.params;
-    
+            
+            const creatorUserType = req.createUserType;
+
+            // Apenas administradores podem excluir outros usuários
+            if (creatorUserType !== 'Administrator') {
+                console.log(req.userType);
+                return res.status(403).json({ error: "Acesso proibido. Apenas administradores podem excluir conteúdo do catalógo." });
+            }
+
             const deletedMovie = await catalogModel.deleteMovie(id);
     
             if (!deletedMovie) {
